Renames related frontstage transform files to element
