+++ tools/codemod/src/v4.0.0/__testfixtures__/element-to-config/Frontstage.output.tsx
