+++ tools/codemod/src/v4.0.0/__testfixtures__/element-to-config/Full.input.tsx
@@ -1,3 +1,7 @@
+/*---------------------------------------------------------------------------------------------
+* Copyright (c) Bentley Systems, Incorporated. All rights reserved.
+* See LICENSE.md in the project root for license terms and full copyright notice.
+*--------------------------------------------------------------------------------------------*/
 import * as React from "react";
 import {
   ContentGroup, Frontstage,
